address some comments (#42)
